requirements.txt
discord.py==2.2.3
langchain==0.0.181
PyYAML==6.0
PyYAML==6.0
